requirements.txt
Django>=5.2.0
djangorestframework
django-cors-headers
django-filter
djoser
djangorestframework-simplejwt
requests
gunicorn
whitenoise
psycopg2-binary
python-decouple
dj-database-url
